requirements.txt
absl-py==0.15.0
aiohttp==3.7.4.post0
aiohttp-cors==0.7.0
aioredis==1.3.1
async-timeout==3.0.1
attrs==21.2.0
blessed==1.19.0
cachetools==4.2.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
cloudpickle==2.0.0
colorful==0.5.4
configparser==5.0.2
cycler==0.10.0
dm-tree==0.1.6
docker-pycreds==0.4.0
filelock==3.3.1
gitdb==4.0.9
GitPython==3.1.24
google-api-core==2.1.1
google-auth==2.3.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
gpustat==1.0.0b1
grpcio==1.41.0
gym==0.21.0
hiredis==2.0.0
idna==3.3
imageio==2.9.0
imageio-ffmpeg==0.4.5
jsonschema==4.1.2
kiwisolver==1.3.2
lz4==3.1.3
Markdown==3.3.4
matplotlib==3.4.2
msgpack==1.0.2
multidict==5.2.0
networkx==2.6.3
numpy==1.21.3
nvidia-ml-py3==7.352.0
oauthlib==3.1.1
opencensus==0.8.0
opencensus-context==0.1.2
pandas==1.3.4
pathtools==0.1.2
Pillow==8.4.0
prometheus-client==0.11.0
promise==2.3
protobuf==3.19.0
psutil==5.8.0
py-spy==0.3.10
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.1
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.1.1
PyYAML==6.0
ray==1.7.1
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.3
scipy==1.7.1
sentry-sdk==1.4.3
shortuuid==1.0.1
six==1.16.0
smmap==5.0.0
subprocess32==3.5.4
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4
termcolor==1.1.0
tifffile==2021.10.12
torch==1.10.0
torchvision==0.11.1
typing-extensions==3.10.0.2
urllib3==1.26.7
wandb==0.12.5
wcwidth==0.2.5
Werkzeug==2.0.2
yarl==1.7.0
yaspin==2.1.0
